Updated the styling of the DogCards component per Asha's last pull request.
