+++ client/src/pages/Home.js
@@ -1,15 +1,45 @@
 import * as React from "react";
+import { useQuery } from "@apollo/client";
+import { GET_ALL_DOGS } from "../utils/queries/dogQueries";
 import DogCards from '../components/DogCards';
 import MainLoggedOut from '../components/MainLoggedOut';
 import Auth from '../utils/auth';
+import Grid from '@mui/material/Grid';
 
 export default function Home() {
+  //Use query to get all dogs from the database
+  const dogData = useQuery(GET_ALL_DOGS);
+  const dogArray = dogData.data?.getAllDogs || [];
+
   return (
     <div className="main-container">
       {Auth.loggedIn()
-        ?
-        <DogCards />
-        :
+        ? (
+          <Grid container spacing={2} justifyContent="center">
+            {dogData.loading ? (
+              <h1>Loading...</h1>
+            ) : (
+              <>
+
+                {/* <h1>Find New Furends!</h1> */}
+                {dogArray.map((item, key) => {
+                  console.log(`I'm hit!`);
+                  return (
+                    <DogCards
+                      key={key}
+                      images={item.images}
+                      name={item.name}
+                      breed={item.breed}
+                      sex={item.sex}
+                      headline={item.headline}
+                      about={item.about}
+                    />
+                  );
+                })}
+              </>
+            )}
+          </Grid>
+        ) :
         <MainLoggedOut />
       }
     </div>
